src/requirements.txt
aiogram==3.20.0.post0
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.1.0
requests==2.32.3
sqlalchemy==2.0.41